refactor into docker base test suite for java connectors
